requirements.txt
colorlog==2.10.0
colorama==0.3.7

## ruamel is recommended over pyyaml because it can preserve comments
## and support newer standard of yaml
ruamel.yaml==0.14.2

## Commandline client
click==6.7
